Reverserer en del flytting av kode
