+++ packages/fakta-beregning/src/components/fellesFaktaForATFLogSN/vurderFaktaBeregningHjelpefunksjoner.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const parseStringToBoolean = (value: string) => value === 'true';
